+++ src/Service/InvoiceGeneratorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FreshAdvance\Invoice\Service;
+
+use FreshAdvance\Invoice\DataType\InvoiceDataInterface;
+
+interface InvoiceGeneratorInterface
+{
+    public function generate(InvoiceDataInterface $invoiceData): void;
+}